core: Add TPM data persistence
TPM devices provide a small storage space to store encryption keys or
other data securely.  Current TPM device support in oVirt doesn't
store TPM data across VM runs, which makes TPM use very limited.

TPM data is stored by swtpm on the local file system of the host
during VM run.  We must transfer the data to Engine after each update
and transfer it back to the host before we start a VM.  This is
achieved by the following additions, implemented by this patch and its
Vdsm counterpart (https://gerrit.ovirt.org/c/vdsm/+/112492 and following):

- An additional optional parameter `_X_tpmdata' is passed to VM.create
API call, containing TPM data to start the VM with.

- Before VM.destroy is called, TPM data is retrieved and stored to the
  database.  It is important to handle retrieval failures properly:
  Not to lose updated TPM data and not to prevent VM destroy forever.
  TPM data is currently expected to be retrieved only for VMs that are
  powered off, not for destroyed VMs after successful live migrations,
  especially not if the data is no longer present on the source host.

- A new verb VM.getExternalData is introduced to retrieve TPM data
  from Vdsm.  This verb is intended to be used also for other
  contingent VM data transfers from the host, such as secure boot
  data.  Since updates of this kind of data are expected to be
  infrequent, the verb asks just for a single piece of data for
  simplicity, rather than performing bulk operations similar to
  e.g. Host.dumpxmls.

TPM data is stored in a new separate database table to avoid
unnecessary retrievals in views and other entities that include all
content of vm_dynamic table.  The table is indexed by VM device id to
ensure data consistency -- if a TPM device is removed, its data is
automatically removed too.  (Future NVRAM data is supposed to be
placed in a different table.)

When TPM is removed from the VM, TPM data is removed too, in order not
to keep and possibly copy no longer needed and sensitive information
around.  To prevent accidental deletion of the data, a confirmation
dialog is displayed when TPM is removed from a VM containing TPM data.

Both this patch and the Vdsm patches are needed for TPM data
persistence.  If any of them is missing, everything will work normally
except for TPM data persistence, TPM data won't be stored.

This patch provides only minimum necessary functionality for TPM data
persistence.  It has the following drawbacks that are going to be
addressed in followup patches:

- Snapshots, with or without memory, don't store TPM data.

- There is no attempt to hide TPM data in logs, in the database or in
  the storage (e.g. in suspended VMs).

- TPM data is retrieved only once the VM is down or is going to get
  powered off.  It should also be checked and retrieved periodically, in
  case the host goes down while the VM is running.

- TPM data retrieval is limited to three repeated attempts.  If they
  all fail, the data is not retrieved and the VM is destroyed. This is
  to prevent blocking VM destroy forever.

Change-Id: I6646f8926cf601b6a044a561f4e7873c6bcfde18
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>