core: skip stopping NBD server for downloading a backup image
When downloading a backup image, QEMU NBD server is used and
not VDSM NBD server.

Therefore the VDSM NBD server is not started so it should not
be stopped.

We currently succeeded to stop the NBD server since this API is
designed to do nothing if NBD server is not running.

Change-Id: Ib2c62c3d245901d1b6c32c6e526623049c7cd2b1
Bug-Url: https://bugzilla.redhat.com/1900540
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>