core: persist backup host UUID
Persist the host UUID that the VM is running on during the backup
under VmBackup in vm_backups table.

The host UUID will be used when the backup will be stopped, this will
save the dependency on a running VM when the backup is stopped.

For example, when a VM powered-off during the backup from the guest, and a
request to stop the backup sent, there will be no host to stop the backup
since the VM is no longer running.

After this change, the persistent host UUID will be used instead of trying
to use the host that runs the VM.

Change-Id: I32b6790372bd839e6f77335100af588e863314ec
Bug-Url: https://bugzilla.redhat.com/1900594
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>