core: block sparsifying qcow volumes
It is possible to trigger virt-sparsify on QCOW volumes - unused space would be
reclaimed. However, from the usability point of view it is problematic since the
sparsification is not reflected on the image size (as there's no defragmentation)
and thus users may think that the operation didn't do anything.

We already block sparsifying disks with snapshots. This patch extends that also
to disks with a single qcow volum because of the reason mentioned above.

Change-Id: Idd2655bb4e6d50562d4596f157f62ef3e34922bd
Bug-Url: https://bugzilla.redhat.com/1792905
Signed-off-by: Arik Hadas <ahadas@redhat.com>