vdsbroker: change the serial to image id
LUNs were not set with serial. This caused to libvirt auto-set it as the
disk alias. We changed it into the LUN ID, but this harms backwards
compatibility. Therefore, we now aligned with non-LUN disks, having the
serial set with the image ID. The only different case is LUN with
pass-through, where we can't set the serial and it will exist as the LUN
ID.

Change-Id: Ib6d38460a459ac724835f12e73f904a1c83929b4
Bug-Url: https://bugzilla.redhat.com/1904774
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>