use tags views in inner queries when needed
This patch verifies that if the generated query is using tags then all
the main entity fields involved in the inner search will use the view
that supports tag rather than the native entity view.

For example, as the bug describes , searching for :
"Vms: status=up and storage=iscsi_L1_vms_sd_2" is generating a query
that is using vms_with_tags , but the inner query uses "vms.status"
instead of "vms_with_tags.status" which is already there, adding a huge
redundant join to the nested query of vms*vms_with_tags.

This is not a special case and the patch handles a solution to all
entities that have a "*_with_tags" corresponding view (users, vms and
hosts)

This patch includes also a new test and modified tests that will perform
better after this fix.

Change-Id: I058f1f8c1e4226adde6f0fce733e28cbbcaa1a95
Bug-Url: https://bugzilla.redhat.com/1581677
Signed-off-by: Eli Mesika <emesika@redhat.com>