core: handle missing virtio-win
When having a windows VM in the cluster and the cluster version is equal
or above 4.4, we support VirtIO-Win RPM. When it's not installed we will
recieve null for the ISO name and the agent version. This caused to an
error, preventing from using and showing ISO from the ISO domain.
Now, we will handle the null values when the RPM isn't installed
allowing ISO domain functionality to continue.

Change-Id: I59724f34865f109dcebe551e3efcb157b6a3c82d
Bug-Url: https://bugzilla.redhat.com/1904947
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>