sso: restore username@profile as user_id
This patch restores username@profile in user_id field of Token Info
Servlet's response. OVN provider actually expects it in this format.

The reason why OVN provider was not updated is the fact that there might
be more similar Token Info Response consumers that would need to be
updated.

Restoring the functionality seems to be the most sane and safe
option.

Change-Id: Iba021aa3a6aaaa94880868534e56e48db8301f8b
Bug-Url: https://bugzilla.redhat.com/1884233
Signed-off-by: Artur Socha <asocha@redhat.com>