sso: better error msg  when user cannot login
This patch fixes a message that is displayed to the user when incorrect
credentials are used. The message is now exactly the same as before
changes in the patch [1]
Additionally, error log statement now handles case when authz name
cannot be resolved - which is legit at that early stage of
authentication.

Change-Id: I9f46679c244bb68279f9e0822da4d270882ca795
Bug-Url: https://bugzilla.redhat.com/1884233
Signed-off-by: Artur Socha <asocha@redhat.com>