vdsbroker: fix NPE on mdev device
MDEV is a host device, cannot be correlated with any of the devices
of the host that the engine knows of. When running a VM with such device
it has a MDEV on it and fails to find it on the host devices the engine
knows of. This results in NullPointerException we need to prevent.

Change-Id: I60626d3fa4be9d34a12969104467af0b94cea5bb
Bug-Url: https://bugzilla.redhat.com/1905417
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>