core: drop checkpoint_xml from vm_checkpoints table
Libvirt doesn't require now the VM domain XML when checkpoint redefines,
so now there is no need to persist the VM checkpoint XML in the database.

The host can now compose the checkpoint XML according to a given backup
configuration (will be implemented in the following patches).

This patch removes the checkpoint_xml from the database and all the code
that handles this field in StartVmBackupCommand.

Change-Id: I6dead518af0f385207a032f1845b4bf3ec3514a2
Bug-Url: https://bugzilla.redhat.com/1901835
Bug-Url: https://bugzilla.redhat.com/1891470
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>