core: redefine checkpoints without checkpoint XML
In the previous patches, the checkpoint_xml column was dropped from
the vm_checkpoints table. This column held the checkpoint XML in order
to use it later when the checkpoint will be redefined for the VM.

Since patch - https://gerrit.ovirt.org/#/c/vdsm/+/112576/,
VDSM supports checkpoint redefinition using the checkpoint XML or
by creating the checkpoint XML on the host using the backup
configuration.

This patch changes checkpoints redefinition and uses the new option to
redefine a checkpoint using the backup configuration.

Change-Id: If8462d5df15fa5e2682116c88ce8e5d4d1030416
Bug-Url: https://bugzilla.redhat.com/1901835
Bug-Url: https://bugzilla.redhat.com/1891470
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>