storage: implement finished tasks cleanup via REST API
Currently, the storage pool manager (SPM) fails to
switch to another host if the SPM has uncleared tasks.
Clearing all the finished tasks enables the SPM switching.
Adding the REST API implementation for that purpose.

For example, to clean all the finished tasks on a
data center with ID `123` send a request like this:

POST /ovirt-engine/api/datacenters/123/cleanfinishedtasks

With a request body like this:

<action/>

Change-Id: I8c36f61a85a1b2975a8dca4058ef7b652a3c04da
Bug-Url: https://bugzilla.redhat.com/1627997
Signed-off-by: Pavel Bar <pbar@redhat.com>