core: introduce fips mode to cluster
When we add hosts to a cluster, we expect the same capatibilities from
them in order to allow normal functionality such as VM migrations. Until
now we allowed having hosts with FIPS and without it within the same
cluster which broke this behavior. We will now detect the first host in
the cluster and set the cluster FIPS accordingly. Hosts incompatible
with the cluster FIPS will become non-operational.

Follow up patches will introduce it to UI and REST-API.

Change-Id: Id077ede5e54af3fbbd30329d9bc81cecee27bce1
Bug-Url: https://bugzilla.redhat.com/1712481
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>