core: fix import OVA on PPC
We used to fail the selection of PPC hosts as proxy for import operations
since they lacked virt-v2v. This still holds for importing from an external
provider (e.g., VMware, XEN) and external OVAs that require to use virt-v2v
but it doesn't hold for importing OVAs that were created by oVirt.

This patch changes the validation of the proxy host in ImportVmFromOva to
enable importing OVAs that were created by oVirt using PPC hosts.

Change-Id: Iea85f413f6eee4eb546b2b7a3122a95eed03d528
Bug-Url: https://bugzilla.redhat.com/1886520
Signed-off-by: Arik Hadas <ahadas@redhat.com>