core: handle hosts after fips update
Once we update the cluster to new FIPS mode, the cluster should change
accordingly. If it was set to `undefinied`, the available host will set
the cluster to the required FIPS mode. The hosts should be in
maintenance while we do the update. In case of race that one or more
hosts are being activated, hosts that are not fulfilling the cluster
criteria should become non-operational.

Change-Id: I1cc6bcd175cbdb61e64d9e079662dbc7e2d28f7d
Bug-Url: https://bugzilla.redhat.com/1712481
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>