core: prevent non-admin users see other users data
This patch fixes a security hole that enables regular users to access
other user data including administrators.
The problem was in the DAO that accesses the users data according to the
user permission, the wrong logic was to get all the user data if any
permission is found for the given user.

This patch modifies the relevant queries in the BLL level to return only
the information that the user allowed to see

CVE-2020-35497

Change-Id: I5130799027ab79f03b4e25c5f2f2ca4150887719
Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=1899938
Signed-off-by: Eli Mesika <emesika@redhat.com>
(cherry picked from commit 40160e6f678d632937a22a8e23370086024f9994)