core: handle single threaded host for auto pinning
The feature is based on suitable hardware but it might not always be
presented. While the users will try to automate the pinning they might
not be aware of that. This patch fixes the gap in that topic about host
having only one thread. We will have the CPU pinning set, but each vCPU
will have only single pCPU pinned instead of two and that pCPU won't be
used for multiple vCPUs.

Change-Id: I87486b994f5d32f42ac62a4b3c907ff2116d6f15
Bug-Url: https://bugzilla.redhat.com/1908757
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>