core: finished tasks cleanup (proper error handling)
When an error is raised during finished tasks cleanup
execution, the error is not properly propagated back to
the caller in order to be displayed to the user.

This patch solves it by using the built-in mechanisms
for error handling.

For example, previously in case of an error, the REST
response showed an empty and useless message:
<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
<fault>
    <detail>[]</detail>
    <reason>Operation Failed</reason>
</fault>

While after the fix the error description is included, i.e.:
<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
<fault>
    <detail>[Network error during communication with the Host.]</detail>
    <reason>Operation Failed</reason>
</fault>

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1910302
Change-Id: I548a70e47024af075dc59b7f85e68f3f84c73fee