core: adjust vm to chipset on change-cluster
When moving a VM that is set with bios-type = CLUSTER-DEFAULT to
a cluster that is set with a different chipset (compared to the
cluster that the VM originally resided in), the VM configuration
may need to be adjusted in order to comply with the new chipset.

This patch applies the logic we currently have for adjusting VMs
to a different chipset also to the above mentioned flow.

Change-Id: I79f8fb3ae7dc0820c6ebeb8ef3cb91d4efee1575
Bug-Url: https://bugzilla.redhat.com/1894454
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>
Signed-off-by: Arik Hadas <ahadas@redhat.com>