core: remove dedicated_vm_for_vds from vm_static
We stored the hosts that a VM is pinned to in two places in the database -
vm_static and vm_host_pinning_map. That's redundant and as we recently
experienced can also be problematic in some scenarios.

This patch removes that from vm_static. In order to be able to get the data
from vm_static along with the host pinning, the vm_static_view is added and
it is now used almost everywhere vm_static was, except for very few places
that need a subset of vm_static that does not include the pinning.

This change doesn't break the VM views since they don't rely on the host
pinning in vm_static already and it should be alright from the performance
perspective since queries of vm_static that are changed to query the
vm_static_view are quite rare.

This patch also remove VmBase#setDedicatedVmForVdsList(Guid) - this utility
method just adds more code to the VM classes and is not really needed.

Change-Id: I1bf25737d982ff1d40200f0f50e274a26c0addd7
Bug-Url: https://bugzilla.redhat.com/1895217
Signed-off-by: Arik Hadas <ahadas@redhat.com>