core+api: fix numa next run update
The default numa tune mode is `interleaved` when not sent by the API
request we need to set it as the default. Otherwise, it might be
nullable which is not expected by the engine. Once we update a next run
configuration for VM NUMA nodes we are clearing the existing DB data and
setting the new one. On multiple calls without shutdown it means we are
losing the VM NUMA data. The patch will preserve that data.

Change-Id: I8fa91aa4ff6e7a055e1b7f440cfe691757476756
Bug-Url: https://bugzilla.redhat.com/1890665
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>