core: add backup scratch disk type to content type
When a backup is taken for a VM, for each disk that
participates in the backup a scratch disk is created.

This patch adds this disk type to the content type
so it will be easy to distinguish between the scratch
that was created during a backup and regular data disks.

Change-Id: Ibb01a9d0b88b9e9cf21e435287a6c395b7afbafd
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>