engine: Add support of `bochs` display type
Adds a new display type `bochs` and enables it for UEFI VMs. The
bochs-display device is a simple display device which is similar to VGA
but without legacy emulation.

The support is enabled for clusters version 4.5 and above.

Bochs is set as default display type for 'Server' VM configuration.

For further implementation details, refer
https://www.ovirt.org/develop/release-management/features/virt/bochs-display.html

Change-Id: If5020b8e9211f8b453356f8d27fbb9964c78b78d
Bug-Url: https://bugzilla.redhat.com/1753645
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>