core: create VM backup scratch disks on shared storage
Until now, when a backup is created for a VM a scratch disk is
created for each disk that participates in the backup.
The scratch disks are created on the local storage of the host
that runs the VM.

The host local storage is limited and if the storage runs out during
the backup due to intensive data written to the VM, the VM will
be paused.

To avoid this situation, the scratch disks will be created on the
same shared storage as the backed-up disk resides. To allow that,
the engine needs to create the scratch disk and prepare it before
starting the backup.

This patch adds the scratch disks creation and preparation before
starting the VM backup.

Change-Id: Idce7c905fb0796193250bfa7c536df43f1880df0
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>