core: remove VM backup scratch disks when the backup finalized
Now, when scratch disks created on the shared storage and
prepared before the VM backup started, the engine should
clean-up those disks when the backup ends or failed.

This patch adds the scratch disks cleanup.
When the backup ended, the engine will teardown the created
scratch disks and remove them as a best effort.

Change-Id: I0d12c9d2cb0adf8841a5a1dbdcea486fb0000390
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>