core: fix remove checkpoint cleanup in case of a backup error
If a failure occurs after the VM backup checkpoint already
created in the DB but before it was set in the VM backup
entity in the command parameters, the checkpoint will not
be removed.

This patch fixes this by using the created checkpoint ID
that was set when the checkpoint being created.

Change-Id: Ic4d8f40988561b4b9ac5e46fbf4745cca272ec2c
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>