core: Bios type handling for templates
Templates do not have to sync its effective
bios type with cluster bios type since such
modifications are made during creating a new
vm from the template. That means we do not have
to change the effective bios type and consequently
update the devices during add or imort a template.

However, the effective bios type could be changed
also by changing the custom bios type of the template
that is why we need to update the effective bios type
and change the the devices in UpdateVmCommand.

Change-Id: I692f391c796c1ac014cde8903e7bcf1f183c8042
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>