core: introduce CreateScratchDiskCommand
When a backup is taken for a VM a scratch disk is added
for each disk that participates in the backup. Until now,
the creation of those disks occurred in the host before the
backup started, and the created disks reside on the host
local storage.

CreateScratchDiskCommand will be used later to create
a scratch disk for a disk that participates in the
VM backup on the same shared storage which holds the
backed-up disk.

The creation of the scratch disks can occur simultaneously
using this command.

Change-Id: I9b87e5e7e63f2b837ac432850dd23ef72d2acf74
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>