core: introduce CreateScratchDisksCommand
CreateScratchDisksCommand will be used later to create
a scratch disk for each disk that participates in the
VM backup on the same shared storage which holds the
backed-up disk.

CreateScratchDisksCommand is using CreateScratchDiskCommand to
create all the scratch disks simultaneously. The command ends
when all the requested scratch disks were created.

Change-Id: Ic9a8b9a7b9c6104711b22bcc4dcedbcbd6bf9570
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>