core: add timeout policy to an image transfer
After the timeout is expired in upload image transfer,
the transfer will become inactive, and the VM will be
left with a disk in locked state.

In upload, we cancel the transfer and delete the disk.
The issue here is download (for example during backup).
If the backup application fails to finalize the transfer
(e.g. crash on backup application side or bug), the transfer
pauses and the disk remains locked.

We cannot change the current behavior since existing users
may depend on it. Adding a timeout policy allow new users to
control the behavior without affecting existing users.

In case the timeout policy is 'cancel', the system will
cancel the transfer and unlock the disk on client
inactivity timeout.

If the transfer policy is 'pause', then, the transfer
will be paused on client inactivity timeout.

The default behaviour is the 'legacy' policy,
which cancels the transfer if the direction is
download, and pauses it if it is upload.

Change-Id: I07c2e178c2d2045d390413c439b5beb741414987
Bug-Url: https://bugzilla.redhat.com/1892676
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>