core: Prevent hibernating VMs with NVDIMMs
Due to a platform problem, hibernating VMs with NVDIMMs is very slow,
see https://bugzilla.redhat.com/1902691.  Until that problem is fixed,
it's better to prevent hibernation of VMs rather than exposing a
functionality that may "never" finish.

This patch blocks the hibernation after pressing "Suspend" button.  It
doesn't disable the button itself because buttons are apparently
disabled only for VMs in certain states and without providing
explanations.

Once the platform bug is fixed, this patch can be reverted and
https://bugzilla.redhat.com/1897906 closed.

Change-Id: Ia0d096ef16e604b0564642badfa1b5919fb7bff9
Bug-Url: https://bugzilla.redhat.com/1912426
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>