core: finished tasks cleanup (partial failure update)
Some modification of the logic for calculating the
audit log partial vs. full failure.

Previously every network error (that resulted in terminating
the cleanup flow) was considered as a "full failure" even if
there were 1 or more tasks that were successfully cleaned-up.
On the other hand other errors were considered as a
"partial failure" even if eventually there were no tasks
that were cleaned-up successfully.

Now the logic makes more sense: if there is at least 1 task that
was successfully cleaned-up and at least 1 task that failed to be
cleaned-up, this is considered a partial failure.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1910302
Change-Id: I92e3b3d6f1d03b31ec7a750093a412e6dc0ad1e4