webadmin: 'Refresh LUN' from VM Disks sub-tab
screenshot: https://imgur.com/a/Rjid65i

'Refresh LUN' operation is now accessible from the menu on the
VM Disks sub-tab.
This button syncs the LUN's size on all hosts connected to the
LUN disk, and updates its size on the VM it's being attached to
while the VM is running.
In case the disk is attached to more than one running VM, all VMs
are being updated.
In case of a not running VM, the disk should be updated once the
VM is up.

Change-Id: I1cdd9b4bcb84e9b5aac4c31636b5a63c73cb2fba
Bug-Url: https://bugzilla.redhat.com/1155275
Signed-off-by: Shani Leviim <sleviim@redhat.com>