core: fix qos to network matching
Creating a map from network ids to QoS objects should take into account
that there may be several networks using the same QoS. Therefore each
network has to be inspected for a matching QoS and not as before this
fix where a network might have been skipped.

Also, comparing QoS Guids should be done with full object equals
comparison. Note that a network may not have a QoS so the equals should
execute on the QoS object.

Change-Id: Ifcc61af1e2830fee904df500c1d1d6999213c352
Bug-Url: https://bugzilla.redhat.com/1921119
Signed-off-by: Eitan Raviv <eraviv@redhat.com>