Change notification config to save 7 days history
We had performance issues when running engine-setup while audit_log
table and event_notification_hist are quiet large

engine-setup has a verification process that insures data consistency by
checking that all ids in dependant entity tables includes only ids in
the main table the foreign key is pointed to.

This is time consuming for large history tables like
event_notification_hist and makes no sense since the insertion to this
table is automatic and not manual beside the fact that we should not
remove history tables unless a configured cleanup time frame passed.

This patch set the notification history configuration to clean
notifications that there sent time is older than 7 days in order to
prevent this table from being to large and affect the overall upgrade
time.

Change-Id: I62d4e81e7b24dfac8290538fba5fa9334299def0
Bug-Url: https://bugzilla.redhat.com/1922200
Signed-off-by: Eli Mesika <emesika@redhat.com>