engine: handle tsx removal gracefully
Since CentOS 8.3, the tsx is disabled by
default. This can cause the hosts to
become non operational. This patch introduces
several small changes to prevent that.

- do not allow host upgrade if the cluster
is set with Secure Intel Skylake Client Family,
Secure Intel Skylake Server Family, Secure Intel
Cascadelake Server Family and the cluster cpu
configuration has not been updated to "noTSX".
- show ui warning when upgrading host with affected
insecure cpu
- show ui warning for the cluster (on cluster list
and on cluster detail) that the cluster cpu configuration
is outdated.
- when checking if host complies with the new configuration,
consider states Up and Preparing for maintenance. For
automatic configuration chagne on engine startup or host
activation, if there is a host in the cluster with any other
state, the configuration wont be updated. For update cluster,
since this is a manual action, only hosts with those states
will be validated and the configuration will be applied
no matter if hosts in other states do not comply.

Change-Id: Iec2034a6fdcf4322d2cce51228966de128df595c
Bug-Url: https://bugzilla.redhat.com/1905158
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>