core: fixed problem with bochs on import
When importing a vm from OVA that has specified
video device, but not the graphics device, the
loading of displays from os repository failed,
because the provided version was not valid.

This patch adds validation to the os repository
if the provided version is valid + it
provides a valid version when the os repository
is called from ovf reader.

Change-Id: I409890b46b521d4ba3b2913758408fef274e2805
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>