core: Host deploy flow need to wait until host reboot finishes
At the end of host deploy flow we executed SshHostReboot commend to
actually reboot a host. The problem is, that SshHostRebootCommand
reboots host synchronously, but then it asynchronously changes host
status to Reboot and wait until it reboot. But for host deploy use case
we need to wait until host really finishes booting and continue only
afterwards.

Change-Id: I46babe84d1a4b4a5c73336b5bb3cde9d393d539d
Bug-Url: https://bugzilla.redhat.com/1917821
Signed-off-by: Martin Perina <mperina@redhat.com>