core: refresh LUN info on all VMs' hosts instead of one
In case of having two hosts that are using the same LUN
(for example, a shared LUN attached to 2 VMs, each one is running
on a different host), and the LUN was refreshed with 'Refresh LUN',
only the host of the first VM found was refreshed using GetDeviceList.
This patch refreshes all the hosts of the plugged VMs.
For non-running VMs (getRunOnVds is null), no online update is required.

Change-Id: Ice4cd54c0f7064cb7b185ca949055ceb90fcd7fb
Signed-off-by: Shani Leviim <sleviim@redhat.com>