webadmin: Remove TpmModel auxiliary class
Let's not attempt to watch TPM field changes and set the value of
interest manually.  While it's error prone because the original value
must be set at all the proper places, it's better aligned with the
rest of the frontend and the risk of omitting the extra statement is
relatively low when doing copy&paste&edit.

Change-Id: I309338967da14f9d39682ec3e73f0082a6b1a858
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>