core: allow mixed incremental backup of RAW and COW disks
Previously it was not possible to create an incremental backup
that includes COW and RAW disks. The reason was that RAW
disks were valid only for a full backup. But now we support the option
to report on the taken backup_mode for each disk (full or incremental)
and we can take the needed backup according to the disk format.

This patch changes the validation for having 'incremental backup enabled'
for all the disks that participate in an incremental backup operation to
occur only if the backup_mode feature isn't supported.

For RAW disks, a full backup will always be taken.

Change-Id: I13ca125dc2b9a0c28c060f3f6bb272a2471851a5
Bug-Url: https://bugzilla.redhat.com/1926888
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>