webadmin: Add TPM deletion confirmation to templates
It doesn't cover the case when a VM is created from a template and the
VM TPM is disabled in the created VM dialog.  There is no TPM data for
the VM yet so it doesn't make sense to check it.  The user should know
what he is doing and can re-create the VM from the template if needed.

Change-Id: Iee339a3faff9b3a1cdf6c185191ec2154c65fc57
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>