engine: remove custom_bios_type from vm_static
Previously, when creating a vm, the bios type was
set according to the cluster value (or was overridden
by custom bios type value). This caused a lot of trouble
cause the bios type is more dependant on the os, devices
and disks than on some global value.

This patch changes this and now when creating a vm, the
bios type of the template is copied. The only exception is
a Blank template, since it is cluster
independent and should not contain any bios type specific
parts. When creating a vm from the Blank template, the cluster
bios type is still set as vm's bios type.

With these modifications, there is no more need to store
a custom value (thus removing from vm_static). Since there
is no further calculation between "custom" and "cluster",
we can simply rename the effective_bios_type to bios_type.

Change-Id: I4685efbd36d81b1e6bd655e8b501935e76fc6f56
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>