Add a way to reset a VM
Sometimes you just want to reset a VM without shutdown/start sequence.
In this commit we add an additional option next to a normal Reboot to just hard Reset the VM.
If your VM is frozen (OOM/Crashed/...) and you quickly want to have it back, it might be faster to just do a Reset.

If there are pending changes on the VM, they will also be applied when doing a reset.

Change-Id: Ic26210d83618748df219cb119de5c2200261d9d8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1927718
Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>