core: use Set instead of List for LUNs
This change is a follow-up for patch
https://gerrit.ovirt.org/c/ovirt-engine/+/113400.

In the current code, the getLunsToUpdateInDb() method returns
a List of LUNs from the relevant host(s) that should be updated
on the DB as part of the "Refresh LUNs" operation.

In case of having two hosts that are using the same LUN
(for example, a shared LUN attached to 2 VMs,
each one is running on a different host), and that LUN was refreshed,
it's being returned twice - once for each host.

To avoid it, instead of adding those LUNs to a List,
they're being added to a Set.
However, the LUN was still added twice to the list, since the
_lunConnections parameter wasn't identical between the results, as its
id was set to null.
Therefore, while creating the set, in case of a null id, the storage
server connection is using dao for getting the right value.

Change-Id: If003244ffb48923374747048692043299ec62066
Signed-off-by: Shani Leviim <sleviim@redhat.com>