core: Don't filter migration destinations on TSC frequency if not requested
When migrations are not restricted only to the hosts with the same TSC
frequency, the TSC frequency check should actually be disabled.  This
has been apparently omitted in the original TSC checking
implementation.  Let's replace the VM type check with a TSC frequency
option check to fix it when looking for hosts matching the TSC
frequency.

Change-Id: Ic354092211c2aab5ebc4d4cfdf09a76e8a992ae7
Bug-Url: https://bugzilla.redhat.com/1821199
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>