engine: do not validate custom cpu
Setting custom cpu is an advanced feature that
can be useful to override specific problems
with system wide cpu configuration or to test
things out. However, the current validation
required the CPU to be exactly the same as configured
in vdc_options table, even though it is possible
to enter any value in the webadmin UI.

This patch removes custom cpu validations - as other
custom fileds, it is user's responsibility to enter
the correct values.

Change-Id: I64fe8798fe8744c38102b01094696002a3aeca0b
Bug-Url: https://bugzilla.redhat.com/1905394
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>