core+api: propagate balloon device on copy device
When creating a VmPool or updating one, we usually use the template to
propagate relevant values. Ballooning device was always enabled, but it
was possible to miss it from the base entity in such cases. This patch
will set the right value by letting copy devices to take it from the
base entity when no other value provided.

Change-Id: I41127b92a7158b57844384347dbb46a32fe771fa
Bug-Url: https://bugzilla.redhat.com/1897532
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>