api: fix balloon settings on add vm
When the user created a VM from template and set specific memory policy
to that VM, without setting ballooning the engine set it to `true` by
default without consuming the template configuration in such case. Now,
the VM will be set to the value given if the user specified the
ballooning value or take it from the template as expected.

Change-Id: I312b5a2943d6caf482c33743027a966bf68abd8f
Bug-Url: https://bugzilla.redhat.com/1923717
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>