core: enable incremental backup by default for 4.5 clusters
Incremental backup can be used now by default in 4.5 clusters.
Hosts that added to a 4.5 cluster fully support all the
incremental backup features so for those clusters incremental
backup can be used by default.

For clusters with versions lower than 4.5, the
'IsIncrementalBackupSupported' configuration value in
the database should be set to 'true' explicitly.

Change-Id: Ieefbae99a5dbb92f097b501420e91e90546b5f10
Bug-Url: https://bugzilla.redhat.com/1909197
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>