core: Add NVRAM data persistence
This patch adds support for NVRAM data persistence, similarly to TPM
data persistence.  We retrieve and store NVRAM data persistently when
secure boot is enabled.

As part of the added functionality, EngineException's thrown from
runVdsCommand are handled and counted as failures.  This wasn't
handled previously and could lead to a situation where a VM cannot be
destroyed.

NVRAM data persistence is disabled up to cluster level 4.5 by default
because it is not supported in Vdsm older than 4.4.5.

Similarly to the current TPM support, this patch doesn't address
snapshots and VM exports/imports.  That functionality will be
implemented in future patches.

Change-Id: I3eff8a25e7f6476ca7b2438aad93618ec5c2a49b
Bug-Url: https://bugzilla.redhat.com/1669178
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>