core: set max num of virtual cpus per architecture
The limit of vCPUs on x86 was changed from 384 to 512 on CENTOS/RHEL 8.3
but for PPC it remains 384.

This patch introduces a limit of vCPUs per architecture.
In order to do that the MaxNumOfVmCpus config value was changed to a Map
and the parsing of the configuration values changed to support that.

For entities that are not per-cluster (e.g., instance types), we take
the highest limit for all architectures.

Change-Id: I4bd6572a76a387df6a86f572a3376ac1b993229a
Bug-Url: https://bugzilla.redhat.com/1928705
Signed-off-by: Arik Hadas <ahadas@redhat.com>