core: cleanup created scratch disks in case of failure
In case of failure to create one of the backup scratch disks
there should be cleanup for the disks that already created.

CreateScratchDisksCommand will return now the created disks map
even in case of failure, the map contains the disks that already
created so it can be used in the cleanup of the backup.

Change-Id: I54a0d0d64beccf1281e9d068c9c34b20c484f5b6
Bug-Url: https://bugzilla.redhat.com/1874483
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>