webadmin: show cpu pinning with auto pin
Until now when the user selected a auto pinning policy that is not
disabled, the cpu pinning string was cleared and disabled. That might
cause confusion to users, for example when having HP VM; When the policy
always selected and the CPU pinning string will always be cleared. This
patch will now show the current CPU pinning string of that VM and only
disable the field to existing VMs. As for new VMs it will be cleared and
disabled.

Change-Id: I0590c893fdfee6d20c52e5c01fc461218a67e0dd
Bug-Url: https://bugzilla.redhat.com/1930282
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>