webadmin: fix cpu list when upgrading cluster
When changing the cluster version, the cpu dropdown
is populated with cpu definitions from vdc_options table
for the selected version. When a cpu is selected,
the selection should be preserved by comparing cpu
flags. However, for Intel Skylake Client/Server
IBRS SSBD MDS Family the auto selection has not
worked properly since the change to noTSX-IBRS model.
This patch changes the selected cpu in the dropdown
to the right values for Intel Skylake Client/Server
IBRS SSBD MDS Family.

Change-Id: I066fd77e2fdabb12f8c5ae4e3c53a88e957d3953
Bug-Url: https://bugzilla.redhat.com/1930733
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>