core: modify a query for retrieving metadata and memory volume disks
In case that metadata or memory dump volume is located on the same
storage domain as the snapshot and one of the volumes being transferred
to another storage domain
GetAllMetadataAndMemoryDisksOfSnapshotsOnDifferentStorageDomainsQuery
returns the first volume as volume to transfer even though it is
located with the snapshot on the same storage domain.

This patch fixes the issue and retrieves only the metadata or memory
dump disks that are raised on a different storage domain than the
snapshot.

Change-Id: Id87a0c62570a768f114c586c4b886604d783df44
Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>