core: Fix size calculation for MBS volumes
Previously the size of the disk wasn't properly rounded to the next GiB
multiple. Previously this didn't fail because the MBS disks were
created directly. However, with the introduction of copying this will
be an issue as the size used is the size of a regular image, so
incorrect rounding will create a target volume that is too small.

Change-Id: Ieb9a1ca9adb430dd6f3d4ed8f2ccdfd18b36d820
Bug-Url: https://bugzilla.redhat.com/1906074
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>